+++ apps/web/src/features/calendar-carousel/model/useWeeklyCarousel.ts
@@ -1,9 +1,20 @@
+/**
+ * @module features/calendar-carousel
+ * @description 주간 캘린더 캐러셀 기능
+ * FSD: features layer - 스와이프로 주 이동 기능
+ */
+
 'use client';
 
 import { useState, useMemo, useEffect, useRef } from 'react';
 import { useSwipeable } from 'react-swipeable';
-import { generateWeeklyDates, subDays, addDays } from '../lib';
-import type { CalendarDate } from '../lib';
+import {
+  generateWeeklyDates,
+  subDays,
+  addDays,
+  SWIPE_THRESHOLD_RATIO,
+} from '@/shared/lib/calendar';
+import type { CalendarDate } from '@/shared/lib/calendar';
 
 interface UseWeeklyCarouselProps {
   dates: CalendarDate[];
@@ -27,7 +38,6 @@ export interface UseWeeklyCarouselReturn {
 }
 
 const SLIDE_WIDTH_REM = 39;
-const SWIPE_THRESHOLD_RATIO = 0.25;
 
 export const useWeeklyCarousel = ({
   dates,
@@ -87,7 +97,6 @@ export const useWeeklyCarousel = ({
     onSwiping: (eventData) => {
       if (!isTransitioning) {
         setIsDragging(true);
-        // disableNext가 true이고 왼쪽으로 스와이프(다음 주)하려는 경우 제한
         const delta = eventData.deltaX;
         const clampedOffset =
           disableNext && delta < 0
@@ -100,7 +109,6 @@ export const useWeeklyCarousel = ({
       const delta = eventData.deltaX;
       setIsDragging(false);
 
-      // disableNext가 true면 왼쪽 스와이프(다음 주) 방지
       if (delta < -SWIPE_THRESHOLD && !disableNext) {
         setIsTransitioning(true);
         pendingActionRef.current = 'left';
