+++ apps/web/src/shared/lib/calendar/constants.ts
@@ -1,0 +1,9 @@
+/**
+ * @module shared/lib/calendar
+ * @description 캘린더 관련 상수
+ * FSD: shared layer
+ */
+
+export const WEEKDAYS = ['월', '화', '수', '목', '금', '토', '일'] as const;
+
+export const SWIPE_THRESHOLD_RATIO = 0.25;
