+++ apps/web/src/shared/lib/calendar/dateUtils.ts
@@ -1,3 +1,9 @@
+/**
+ * @module shared/lib/calendar
+ * @description 순수 날짜 계산 및 검증 유틸리티 함수
+ * FSD: shared layer - 비즈니스 로직 없는 재사용 가능한 함수들
+ */
+
 import {
   isSameDay,
   isToday as dateFnsIsToday,
@@ -12,10 +18,18 @@ import {
   startOfWeek,
 } from 'date-fns';
 
-// Re-export date-fns functions
+// ==================== Re-exports ====================
 export { addMonths, subMonths, addDays, subDays };
 export { isSameDay as isSameDate };
 export { dateFnsIsToday as isToday };
+
+// ==================== Types ====================
+export interface CalendarDate {
+  date: Date;
+  isCurrentMonth: boolean;
+}
+
+// ==================== Date Validation ====================
 
 /**
  * 주어진 날짜가 오늘 이후인지 확인 (미래 날짜인지)
@@ -51,13 +65,14 @@ export const isCurrentWeek = (baseDate: Date): boolean => {
   const weekStart = startOfWeek(baseDate, { weekStartsOn: 1 });
   const weekEnd = addDays(weekStart, 6);
 
-  // 오늘이 해당 주의 시작일과 종료일 사이에 있는지 확인
   const todayTime = today.setHours(0, 0, 0, 0);
   const weekStartTime = weekStart.setHours(0, 0, 0, 0);
   const weekEndTime = weekEnd.setHours(23, 59, 59, 999);
 
   return todayTime >= weekStartTime && todayTime <= weekEndTime;
 };
+
+// ==================== Date Formatting ====================
 
 /**
  * 년/월을 "YYYY년 M월" 형식으로 포맷
@@ -66,13 +81,12 @@ export const formatYearMonth = (date: Date): string => {
   return format(date, 'yyyy년 M월');
 };
 
-export interface CalendarDate {
-  date: Date;
-  isCurrentMonth: boolean;
-}
+// ==================== Date Generation ====================
 
 /**
- * 캘린더에 표시할 날짜 배열 생성 (이전 달, 현재 달, 다음 달 포함)
+ * 월간 캘린더에 표시할 날짜 배열 생성 (이전 달, 현재 달, 다음 달 포함)
+ * @param currentDate 기준 날짜
+ * @returns 캘린더 그리드에 표시할 날짜 배열 (보통 35~42개)
  */
 export const generateCalendarDates = (currentDate: Date): CalendarDate[] => {
   const firstDay = startOfMonth(currentDate);
@@ -120,14 +134,13 @@ export const generateCalendarDates = (currentDate: Date): CalendarDate[] => {
 /**
  * 주간 캘린더에 표시할 날짜 배열 생성 (월~일 7일)
  * @param baseDate 기준 날짜 (이 날짜가 포함된 주의 월~일을 반환)
+ * @returns 7일치 날짜 배열
  */
 export const generateWeeklyDates = (baseDate: Date): CalendarDate[] => {
-  // 월요일을 주의 시작으로 설정
-  const weekStart = startOfWeek(baseDate, { weekStartsOn: 1 });
+  const weekStart = startOfWeek(baseDate, { weekStartsOn: 1 }); // 월요일 시작
 
   const weekDates: CalendarDate[] = [];
 
-  // 월요일부터 일요일까지 7일 생성
   for (let i = 0; i < 7; i++) {
     const date = addDays(weekStart, i);
     weekDates.push({
