+++ apps/web/src/shared/lib/calendar/index.ts
@@ -1,0 +1,9 @@
+/**
+ * @module shared/lib/calendar
+ * @description 캘린더 관련 순수 유틸리티 함수 모듈
+ * FSD: shared layer
+ */
+
+export * from './dateUtils';
+export * from './constants';
+export type { CalendarDate } from './dateUtils';
