+++ apps/web/src/shared/ui/calendar/index.ts
@@ -1,6 +1,14 @@
+/**
+ * @module shared/ui/calendar
+ * @description 순수 캘린더 UI 컴포넌트들
+ * FSD: shared layer - 재사용 가능한 UI 컴포넌트만
+ */
+
+// Basic UI Components
+export { DateCell } from './DateCell';
 export { CalendarGrid } from './CalendarGrid';
 export { CalendarHeader } from './CalendarHeader';
-export { DateCell, type DateCellProps } from './DateCell';
+export { WeeklyDateGrid } from './ui/WeeklyDateGrid';
+export { WeeklyHeader } from './ui/WeeklyHeader';
 
-export { formatYearMonth, isSameDate, isToday } from './lib';
-export type { CalendarDate } from './lib';
+export type { DateCellProps } from './DateCell';
