--- apps/web/src/shared/ui/calendar/lib/index.ts
@@ -1,16 +1,0 @@
-export {
-  formatYearMonth,
-  generateCalendarDates,
-  generateWeeklyDates,
-  isSameDate,
-  isToday,
-  addMonths,
-  subMonths,
-  addDays,
-  subDays,
-  isAfterCurrentMonth,
-  isCurrentWeek,
-  isAfterToday,
-} from './calendarUtils';
-
-export type { CalendarDate } from './calendarUtils';
