--- apps/web/src/shared/ui/calendar/model/constants.ts
@@ -1,1 +1,0 @@
-export const WEEKDAYS = ['월', '화', '수', '목', '금', '토', '일'];
