--- apps/web/src/shared/ui/calendar/model/index.ts
@@ -1,3 +1,0 @@
-export { useWeeklyCarousel } from './useWeeklyCarousel';
-export { useMonthlyCarousel } from './useMonthlyCarousel';
-export { WEEKDAYS } from './constants';
