+++ apps/web/src/widgets/calendar/index.ts
@@ -1,1 +1,8 @@
-export { Calendar, type CalendarProps } from './ui/Calendar';
+/**
+ * @module widgets/calendar
+ * @description 캘린더 위젯 모듈
+ * FSD: widgets layer - 조합된 캘린더 컴포넌트
+ */
+
+export { Calendar } from './ui/Calendar';
+export type { CalendarProps } from './ui/Calendar';
